Fix race condition when flake checks run in parallel (#20294)
The Flake checks run in parallel and when you had an image which
required rebuild, it performed additional check on whether the
image needs build or "pull+build". When it was run in parallel
a temporary file containing hash of the remote image could be
overwritten and emptied while another process was reading it
which resulted in error when running flake command.

This has been changed - the files are now stored in a temporary
files - unique to each of the processes running in parallel and
the file in question is moved as an atomic operation so it will
never become empty.

(cherry picked from commit 0b7734ee36670a70363b5866f7a064a5a0d67be8)