Fix generation of "Status provider" issue (#20621)
The script for generating issue for "Provider status" and release
process did not work well when only subset of providers were released.
The issue was generated including some already released packages
even if they were not released in recent batch of providers (if there
was not even a doc change since last release, the package was considered
as being released again).

This PR fixes it by adding a flag that only considers packages that
are present in dist folder (which matches the process of release
manager)

The process has also been updated with more accurate description of
the steps to take - including manual execution of the script rather
than using Breeze (Breeze is not neede for this script).

(cherry picked from commit d823cf7406092cf5b9b7b6df46738cd95a08c485)